LPS-196641 feature-flag-web: (prep) extract FeatureFlagBagProvider as an interface
This is in preparation to turn FeatureFlagBagProviderImpl into an AOPService
so it can be annotated with @Clusterable.
